requirements.txt
crewai>=0.28
langchain-openai>=0.1.0
langchain-community>=0.0.30
langchain-core>=0.2.0
chromadb>=0.5
tiktoken
pdfplumber
python-dotenv
reportlab         
pytest
ruff
black
fastapi
uvicorn
python-multipart
openai>=1.14
google-cloud-vision
google-cloud-storage
aiohttp
requests
python-dotenv
pdfminer.six
mammoth
html-to-docx

